refactor: modularize agent-rules; document inventory vs policy
Split the agent-rules generator into focused modules so hand-written
policy is navigable without generating honesty prose from schema.

- inventory = schema::agent_operations_catalogue (unchanged source)
- packaging = agent_packaging (name map, core tools, explore, honesty)
- policy / surfaces / workflows / reference = curated agent guidance
- Module docs + AGENTS.md structure note the inventory/policy split
